Adds a local (non caching) WebServer
This adds a new dependency 'mime-types' - but that one's already
in the node_modules, so no new `npm install` required. (I
specifically chose this module because of no new dependency needed).

The WebServer can be disabled in the settings and every file needs
to be specifically defined for imports, so no random file access is
allowed. If you are paranoid, you can disable the WebServer, but
some things may break.