Update SettingsOptionFilebox to allow configuration and check for values
The filters and properties for the file dialog aren't hard-wired anymore - they're now read from a property, `dialog` which is passed as an object. In addition to that, the callback also checks for fileName meaning there won't be an error in the console log if someone cancels without selecting a file. In addition to the check that will prevent errors, it will also be compared to a `dialog.required`. If that prop is truthy, the user is forced to select a file.